requirements.txt
Django==5.1.1
psycopg2
pgvector
Pillow==9.5.0
sentence-transformers
djangorestframework
drf-spectacular
google-cloud-aiplatform>=1.29.0 
google-cloud-storage
